[Pipelining] Add debug logs for batch p2p ops
Also include 'dtype' in the existing debug print helper.

Note (BC): technically this replaces a currently 'public' api with a new
private one.  The original public one was only landed recently and has
not been released yet, and should not have any adoption.  The api should
be internal, since we may want to change it and it is intended as a
helper method.

ghstack-source-id: 4a396a3a2fbadcb4ef6baa4b4c9ab5b5723c9a07
Pull Request resolved: https://github.com/pytorch/pytorch/pull/126539